Refactor supergraph resolution to make it more type-safe.
Compose a single time at startup instead of adding/removing subgraphs from subgraph watchers.
