+++ src/command/dev/protocol/follower/mod.rs
@@ -1,7 +1,7 @@
 mod message;
 mod messenger;
 
-pub(crate) use message::SubgraphMessage;
+pub(crate) use message::SubgraphUpdated;
 pub(crate) use messenger::SubgraphWatcherMessenger;
 
 // TODO: switch off of crossbeam
@@ -9,8 +9,8 @@ use crossbeam_channel::{bounded, Receiver, Sender};
 
 #[derive(Debug, Clone)]
 pub struct SubgraphMessageChannel {
-    pub sender: Sender<SubgraphMessage>,
-    pub receiver: Receiver<SubgraphMessage>,
+    pub sender: Sender<SubgraphUpdated>,
+    pub receiver: Receiver<SubgraphUpdated>,
 }
 
 impl SubgraphMessageChannel {
