+++ src/federation.rs
@@ -1,0 +1,5 @@
+//! All things dealing with Apollo Federation, like Composition.
+
+#[cfg(feature = "composition-js")]
+mod composer;
+pub(crate) mod supergraph_config;
